requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
git+https://github.com/openai/CLIP.git
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
faiss-cpu==1.7.4
pydantic==2.5.0
aiofiles==23.2.0
